frontend/ renaming styles and added dirs pages and modals in styles
