+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { Routes, Route, useLocation } from 'react-router-dom';
 import HomePage from './pages/HomePage';
 import UserPage from './pages/DashboardPage';
 import RoomPage from './pages/RoomPage';
-import NotFound from './pages/NotFound';
+import NotFound from './pages/NotFoundPage';
 import { DefaultBackground, DarkBackground } from './components/Background/Backgrounds';
 import './styles/App.module.css';
 import ProtectedRoute from './components/ProtectedRoute';
