+++ frontend/src/styles/pages/RoomPage.module.css
